tests/capi: dump samples in luaL_loadbuffer_proto_test
The test luaL_loadbuffer_proto_test has a large corpus of protobuf
grammars and it is convenient to use it for creating a corpus of
Lua programs. The patch adds this feature. If the directory path
is defined in an environment variable LPM_DUMP_NATIVE_INPUT, then
Lua chunk will be saved to that directory and path to the file
will be printed to stdout. If LPM_DUMP_NATIVE_INPUT is not a real
path Lua chunk will be printed to stdout (old behaviour).

The patch also caches a variable with
`getenv("LPM_DUMP_NATIVE_INPUT")` and then variable is used in
a hot function.