cmake: drop version in target names
Using a branch name with slashes instead of a commit hash causes
problems when configuring CMake because the name cannot contain
slashes. Also, there is not much sense in adding the version to
the target name. So the patch removes a version of the runtime
from the target name.

Needed for google/oss-fuzz#14656